Cab fare Prediction
We are a cab rental start-up company. We have successfully run the pilot project and now want to launch your cab service across the country. We have collected the Historical data from your pilot project and now have a requirement to apply analytics for fare prediction. Now we need to design a system that predicts the fare amount for a cab ride in the city.